requirements.txt
Django==2.0.5
psycopg2-binary==2.7.4
django-mptt==0.9.0
celery==4.1.1
